+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## UNRELEASE
 
+## 2.11.2
+
+- **Fix** : Fixed Paypal not applying discounts
+- **Fix** : Fixed Apple Pay disappears
+- **Fix** : Fixed min/max not filtering card networks
+- **Fix** : Fixed error on Hosted Fields when autocomplete
+
 ## 2.11.1
 
 - **Fix** : fix Mbway phone number issue